extract config component
